fix(security): resolve the 16 open CodeQL alerts
Code (3, the real ones):
- java/unreleased-lock: EffectJournal.forceTo could leave syncLock held if the
  post-fsync watermark advance threw; restructured to a try/finally that always
  unlocks and advances the watermark ONLY on a successful flush (never claims
  durability after an fsync failure).
- java/implicit-cast-in-compound-assignment: TeleportSaga.elapsedTicks was int
  but += a long (CHECK_PERIOD_TICKS, needed long for the scheduler API); widened
  the field to long to drop the narrowing cast.
- java/log-injection: UpdateChecker logged the remote-supplied version/download
  URL verbatim; now collapses CR/LF so a hostile update endpoint can't forge log
  lines.

Actions (6): pinned every third-party action (gradle/actions/setup-gradle +
dependency-submission, softprops/action-gh-release, dependabot/fetch-metadata) to
full commit SHAs with a version comment — no more mutable-tag supply-chain risk;
Dependabot keeps the pins current.

SQL-in-tests (7, all java/concatenated-sql-query): every one is in test code
building SQL from hardcoded in-test fixtures — trusted data, never shipped. Added
.github/codeql/codeql-config.yml scoping the scan to production code (excludes
**/src/test/**, probe/, scripts/), so these auto-resolve and stay out of the
signal. security-extended suite centralized into that config.

Full build + all gates + tests green.

+++ core/src/main/java/dev/fablemc/factions/core/session/TeleportSaga.java
@@ -238,7 +238,7 @@ public final class TeleportSaga implements Travel {
         private final int startX;
         private final int startY;
         private final int startZ;
-        private int elapsedTicks;
+        private long elapsedTicks;   // long: += CHECK_PERIOD_TICKS (long) with no narrowing cast
         private boolean settled;   // single-shot: the pre-teleport money outcome has been decided
 
         Warmup(Player player, PlayerSession session, Location dest, MessageKey doneKey,
